backend/requirements.txt
fastapi==0.109.0
'uvicorn[standard]'
python-multipart
supabase
python-dotenv
openai